# Add partial evaluator to inline before SSA pass