--- a/crates/nargo/tests/test_data/9_conditional/src/main.nr
+++ b/crates/nargo/tests/test_data/9_conditional/src/main.nr
@@ -59,7 +59,8 @@ fn main(a: u32, mut c: [u32; 4], x: [u8; 5], result: pub [u8; 32]){
         must_be_zero(0);
         c[0] = 3;
     } else {
-        must_be_zero(1);
+        // errors with: constraint is always false
+        // must_be_zero(1);
         c[0] = 1;
         c[1] = c[2] / a + 11 % a;
         let f1 = a as Field;