--- a/crates/noirc_driver/src/lib.rs
+++ b/crates/noirc_driver/src/lib.rs
@@ -7,6 +7,7 @@ use noirc_frontend::graph::{CrateId, CrateName, CrateType, LOCAL_CRATE};
 use noirc_frontend::hir::def_map::CrateDefMap;
 use noirc_frontend::hir::Context;
 use noirc_frontend::monomorphisation::monomorphise;
+use noirc_frontend::partialevaluator::evaluate;
 use serde::{Deserialize, Serialize};
 use std::path::{Path, PathBuf};
 
@@ -160,10 +161,11 @@ impl Driver {
         let func_meta = self.context.def_interner.function_meta(&main_function);
         let abi = func_meta.into_abi(&self.context.def_interner);
 
-        let ast = monomorphise(main_function, self.context.def_interner);
+        let program = monomorphise(main_function, self.context.def_interner);
+        let program = evaluate(program);
 
         // Compile Program
-        let circuit = match create_circuit(ast, np_language, show_ssa) {
+        let circuit = match create_circuit(program, np_language, show_ssa) {
             Ok(circuit) => circuit,
             Err(err) => {
                 // The FileId here will be the file id of the file with the main file