--- a/crates/noirc_frontend/src/monomorphisation/ast.rs
+++ b/crates/noirc_frontend/src/monomorphisation/ast.rs
@@ -27,6 +27,77 @@ pub enum Expression {
     Semi(Box<Expression>),
 }
 
+impl Expression {
+    pub fn has_side_effects(&self) -> bool {
+        match self {
+            Expression::Block(exprs) => exprs.iter().any(|expr| expr.has_side_effects()),
+            Expression::Semi(expr) => expr.has_side_effects(),
+            Expression::Literal(Literal::Array(array)) => {
+                array.contents.iter().any(|elem| elem.has_side_effects())
+            }
+
+            Expression::Literal(_) => false,
+            Expression::Ident(_) => false,
+            Expression::Unary(_) => false,
+            Expression::Binary(_) => false,
+            Expression::Index(_) => false,
+            Expression::Cast(_) => false,
+            Expression::Tuple(_) => false,
+            Expression::ExtractTupleField(_, _) => false,
+            Expression::CallBuiltin(_) => false,
+            Expression::CallLowLevel(_) => false,
+
+            Expression::For(_) => unreachable!(),
+            Expression::Call(_) => unreachable!(),
+
+            Expression::If(_) => true,
+            Expression::Let(_) => true,
+            Expression::Constrain(_, _) => true,
+            Expression::Assign(_) => true,
+        }
+    }
+
+    pub fn contains_variables(&self) -> bool {
+        match self {
+            Expression::Ident(_) => true,
+            Expression::Literal(Literal::Array(array)) => {
+                array.contents.iter().any(|elem| elem.contains_variables())
+            }
+            Expression::Literal(_) => false,
+            Expression::Block(exprs) => exprs.iter().any(|expr| expr.contains_variables()),
+            Expression::Unary(unary) => unary.rhs.contains_variables(),
+            Expression::Binary(binary) => {
+                binary.lhs.contains_variables() || binary.rhs.contains_variables()
+            }
+            Expression::Index(index) => {
+                index.collection.contains_variables() || index.index.contains_variables()
+            }
+            Expression::Cast(cast) => cast.lhs.contains_variables(),
+            Expression::For(for_loop) => {
+                for_loop.start_range.contains_variables() || for_loop.end_range.contains_variables()
+            }
+            Expression::If(if_expr) => {
+                if_expr.condition.contains_variables()
+                    || if_expr.consequence.contains_variables()
+                    || if_expr.alternative.as_ref().map_or(false, |alt| alt.contains_variables())
+            }
+            Expression::Tuple(fields) => fields.iter().any(|field| field.contains_variables()),
+            Expression::ExtractTupleField(expr, _) => expr.contains_variables(),
+            Expression::Call(call) => call.arguments.iter().any(|arg| arg.contains_variables()),
+            Expression::CallBuiltin(call) => {
+                call.arguments.iter().any(|arg| arg.contains_variables())
+            }
+            Expression::CallLowLevel(call) => {
+                call.arguments.iter().any(|arg| arg.contains_variables())
+            }
+            Expression::Let(let_expr) => let_expr.expression.contains_variables(),
+            Expression::Constrain(expr, _) => expr.contains_variables(),
+            Expression::Assign(assign) => assign.expression.contains_variables(),
+            Expression::Semi(expr) => expr.contains_variables(),
+        }
+    }
+}
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
 pub struct DefinitionId(pub u32);
 
@@ -37,6 +108,7 @@ pub struct FuncId(pub u32);
 pub struct Ident {
     pub location: Option<Location>,
     pub id: DefinitionId,
+    pub mutable: bool,
     pub name: String,
     pub typ: Type,
 }
@@ -82,6 +154,7 @@ pub struct If {
     pub condition: Box<Expression>,
     pub consequence: Box<Expression>,
     pub alternative: Option<Box<Expression>>,
+    pub typ: Type,
 }
 
 #[derive(Debug, Clone)]
@@ -124,6 +197,7 @@ pub struct Index {
 #[derive(Debug, Clone)]
 pub struct Let {
     pub id: DefinitionId,
+    pub mutable: bool,
     pub name: String,
     pub expression: Box<Expression>,
 }