--- a/crates/noirc_frontend/src/monomorphisation/printer.rs
+++ b/crates/noirc_frontend/src/monomorphisation/printer.rs
@@ -98,6 +98,7 @@ impl AstPrinter {
             super::ast::Literal::Integer(x, _) => write!(f, "{}", x),
             super::ast::Literal::Bool(x) => write!(f, "{}", x),
             super::ast::Literal::Str(s) => write!(f, "{}", s),
+            super::ast::Literal::Unit => write!(f, "()"),
         }
     }