… crates/nargo/tests/prove_and_verify.rs
… crates/nargo/tests/test_data/6_array/src/main.nr