# Related issue(s)

Resolves #461

Related to #374

# Description

## Summary of changes

Adds a new pass to the compiler after monomorphisation to inline all functions, unroll for loops, and evaluate comptime expressions. This means we could technically make SSA a bit smaller by removing these things, but this PR avoids doing this to change less code. There can be a later PR. This makes some features easier to add. As an example, for-loops returning arrays has also been implemented in this PR.

As an example

## Test additions / changes

None yet; I will update with a test for the arrays returned by for loops.

# Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/) and/or `cargo fmt` with default settings.
- [x] I have [linked](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue) this PR to the issue(s) that it resolves.
- [x] I have reviewed the changes on GitHub, line by line.
- [x] I have ensured all changes are covered in the description.