@jfecher 2022-11-14
What is the program you're testing with?
I'm aware of some low hanging fruit optimizations-wise that can be fixed rather easily and can test it versus your input