@guipublic 2022-11-14
> What is the program you're testing with? I'm aware of some low hanging fruit optimizations-wise that can be fixed rather easily and can test it versus your input

I don't think we should spend time on this, even if we re-implement the optimisations done in SSA, it is still not the correct place for inlining and unrolling. This will mess with further static analysis, handling of non-const in loop ranges, return statements, recursion and even custom gates/lookup tables.
For the record, the program I tried is this:
```
use dep::std;
fn main(x: Field, y: Field, z: Field) {
        let _a = x+y+z;
       let mut h1 = std::hash::mimc_bn254([x,y]) as u32;
       for i in 0..100 {
        if x as u32 +1000<5 {
          h1 = std::hash::mimc_bn254([x+i,y]) as u32;
        } else 
        {
        h1 = std::hash::mimc_bn254([x+500,y]) as u32;
        }
       };
     
   constrain h1>1;
}
```