@jfecher 2022-11-15
After timing this some more, it seems the majority of the extra time comes from `optim::full_cse`:
```
PR #462 + minor optimization

Monomorphisation/frontend done: 244.538666ms
Partial evaluation done: 1.02463475s
ssa codegen main done: 4.378943s
  function::inline_all done: 3.416µs
  block::compute_dom done: 54.25µs
  optim::full_cse done: 30.3027525s
  flatten::unroll_tree done: 1.409702833s
  decision tree.reduce done: 6.011341291s
  inline::inline_tree done: 2.997692041s
  block::merge_path done: 3.163667916s
  optim::cse done: 3.160972666s
  integer::overflow_strategy done: 15.337554666s
ssa opts done, total: 62.383857166s
acir gen done: 3.336528291s
total gates: ~12000

--------------------------------
master

Monomorphisation/frontend done: 252.326583ms
ssa codegen main done: 45.760041ms
  function::inline_all done: 1.633393125s
  block::compute_dom done: 28.666µs
  optim::full_cse done: 638.75µs
  flatten::unroll_tree done: 915.583µs
  decision tree.reduce done: 1.97975ms
  inline::inline_tree done: 2.54637325s
  block::merge_path done: 3.075378958s
  optim::cse done: 850.837125ms
  integer::overflow_strategy done: 4.612838541s
ssa opts done, total: 12.722489791s
acir gen done: 1.36219525s
total gates: 6666
```
Although even SSA passes such as integer::overflow_strategy which should occur after all functions are inlined and loops are unrolled normally still take longer.

I think it can still be worth investigating this further since this PR allows changes to the language - like first class functions - more easily than editing the SSA pass and managing its internal state.