@jfecher 2022-11-16
It seems the gate count difference was because the PR was not updated with master. After updating and resolving merge conflicts it actually produces fewer gates than master:

```
jf/partial-evaluator> nargo gates
ACIR gates generated : 6655
Gates successfully counted

Total gates: 11662

By type:
arithmetic: 11631
range: 19
split: 2
truncate: 10

master> nargo gates
ACIR gates generated : 6666
Gates successfully counted

Total gates: 15257

By type:
arithmetic: 15226
range: 19
split: 2
truncate: 10
```