@jfecher 2022-11-17
Closing this PR as it is used by the higher-order-functions PR anyway and the later PR is somewhat more efficient.